+++ src/KsWare.Presentation.ViewFramework/AttachedBehavior/(Xaml.Behavior)/TreeView.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 using KsWare.Presentation.ViewModelFramework;
 
 namespace KsWare.Presentation.ViewFramework.AttachedBehavior {
@@ -10,19 +10,18 @@ namespace KsWare.Presentation.ViewFramework.AttachedBehavior {
 	/// Class BindableSelectedItemBehaviorV40.
 	/// </summary>
 	/// <example>
-	/// <remarks> Requires NuGet System.Windows.Interactivity.WPF 2.0.20525 </remarks>
+	/// <remarks> Requires NuGet Microsoft.Xaml.Behaviors.WPF 1.0 </remarks>
 	/// <code>
-	///		xmlns:e="http://schemas.microsoft.com/expression/2010/interactivity"
+	///		xmlns:e="http://schemas.microsoft.com/xaml/behaviors"
 	/// 
 	///		&lt;e:Interaction.Behaviors&gt;
 	///		&lt;behaviors:BindableSelectedItemBehavior SelectedItem="{Binding SelectedAction.Target, Mode=TwoWay}" /&gt;
 	///		&lt;/e:Interaction.Behaviors&gt;
 	/// </code>
 	/// </example>
-//	[Obsolete(@"Requieres System.Windows.Interactivity.dll (Microsoft SDKs\Expression\Blend\.NETFramework\v4.0)")]
 	public class BindableSelectedItemBehaviorV40 : Behavior<TreeView> {
 		/*
-			Reference: System.Windows.Interactivity (C:\Program Files (x86)\Microsoft SDKs\Expression\Blend\.NETFramework\v4.0\Libraries\System.Windows.Interactivity.dll)
+			Reference: Microsoft.Xaml.Behaviors.dll (NuGet: Microsoft.Xaml.Behaviors.WPF)
 		 
 				<e:Interaction.Behaviors>
 					<behaviors:BindableSelectedItemBehavior SelectedItem="{Binding SelectedAction.Target, Mode=TwoWay}" />