+++ src/KsWare.Presentation.ViewFramework/AttachedBehavior/TreeView.cs
@@ -1,9 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Interactivity;
 using System.Windows.Media;
-using KsWare.Presentation.ViewModelFramework;
 
 namespace KsWare.Presentation.ViewFramework.AttachedBehavior {