+++ src/KsWare.Presentation.ViewFramework/Behaviors.DragDrop/FrameworkElementDropBehavior.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Windows;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 using KsWare.Presentation.ViewFramework.Behaviors.DragDrop;
 
 namespace KsWare.Presentation.ViewFramework.Behaviors {