+++ src/KsWare.Presentation.ViewFramework/Behaviors/(Interactivity.Behavior)/CustomWindowBehavior.cs
@@ -6,7 +6,7 @@ using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 using System.Windows.Interop;
 using Point = System.Windows.Point;
 using System.Drawing;