+++ src/KsWare.Presentation.ViewFramework/Behaviors/(Interactivity.Behavior)/WindowChromeBehavior.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 using System.Windows.Interop;
 using Point = System.Windows.Point;