+++ ChangeLog.md
@@ -1,10 +1,11 @@
 # 0.18
-
+ - Use nuget Microsoft.Xaml.Behaviors.Wpf instead the System.Windows.Interactivity
+   - Read more: https://devblogs.microsoft.com/dotnet/open-sourcing-xaml-behaviors-for-wpf/
+- remove nuget System.Windows.Interactivity.WPF-2.0.20525
 # 0.18.36 (2019-09-03)
 - use nuget KsWare.Presentation.Converters
 - use nuget KsWare.Presentation.ViewFramework.Common
 - change namespace of some converters (KsWare.Presentation.Converters)
-
 # 0.18.22
 - add ViewModelStylesResources
 - add UIElementControllerVM<T>